Updates to Subscription#index view page
This commit fixes the display of subscriptions for guest users, as it
was calling id on nil before.

It also includes a new Subscription scope :post_cart that filters out
cart subscriptions for the Subscription#index view because they're not
valid subscriptions yet. It updates the SubscriptionsController#index
action to use the new scope.